Org change screen

+++ src/main/java/io/synlabs/synvision/service/OrgService.java
@@ -2,6 +2,8 @@ package io.synlabs.synvision.service;
 
 import io.synlabs.synvision.entity.core.Org;
 import io.synlabs.synvision.entity.core.SynVisionUser;
+import io.synlabs.synvision.ex.FileStorageException;
+import io.synlabs.synvision.ex.NotFoundException;
 import io.synlabs.synvision.jpa.OrgRepository;
 import io.synlabs.synvision.views.OrgRequest;
 import io.synlabs.synvision.views.OrgResponse;
@@ -15,9 +17,9 @@ import org.springframework.web.multipart.MultipartFile;
 
 import java.io.File;
 import java.io.IOException;
+import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.util.Objects;
 import java.util.Optional;
 
 @Service
@@ -39,26 +41,34 @@ public class OrgService extends BaseService {
 
     public void saveOrgDetails(OrgRequest request, MultipartFile logoFileMultipart) {
         try {
-            String logoFolder = uploadDirPath + "orgLogo/";
-            File logoFile = new File(logoFolder + logoFileMultipart.getOriginalFilename());
-            FileUtils.copyInputStreamToFile(logoFileMultipart.getInputStream(), logoFile);
+            Path logoFolder = Paths.get(uploadDirPath, "orgLogo");
+            Files.createDirectory(logoFolder);
+
+            if (logoFileMultipart != null) {
+                File logoFile = Paths.get(logoFolder.toString(), logoFileMultipart.getOriginalFilename()).toFile();
+                FileUtils.copyInputStreamToFile(logoFileMultipart.getInputStream(), logoFile);
+            }
 
             if (request.getId() != null && request.getId() != 0) {
                 Optional<Org> opOrg = orgRepository.findById(request.getId());
                 if (opOrg.isPresent()) {
                     Org org = opOrg.get();
 
-                    File prevLogo = new File(logoFolder + org.getLogoFileName());
-                    if (prevLogo.exists()) {
-                        if (prevLogo.delete()) {
-                            logger.info("Previous logo file deleted");
-                        } else {
-                            logger.error("Couldn't delete previous logo file");
+                    if (!request.getLogoFileName().equals(org.getLogoFileName())) {
+                        File prevLogo = Paths.get(logoFolder.toString(), org.getLogoFileName()).toFile();
+                        if (prevLogo.exists()) {
+                            if (prevLogo.delete()) {
+                                logger.info("Previous logo file deleted");
+                            } else {
+                                logger.error("Couldn't delete previous logo file");
+                            }
                         }
                     }
 
                     org.update(request);
                     orgRepository.saveAndFlush(org);
+                } else {
+                    throw new NotFoundException("Couldn't find org with id : " + request.getId());
                 }
             } else {
                 orgRepository.saveAndFlush(new Org(request));
